Move db function outside models
