+++ udata/tests/test_model.py
@@ -7,7 +7,8 @@ from mongoengine.errors import ValidationError
 from mongoengine.fields import BaseField
 
 from udata.settings import Defaults
-from udata.models import db, Dataset, validate_config, build_test_config
+from udata.models import Dataset
+from udata.mongo import db, validate_config, build_test_config
 from udata.errors import ConfigError
 from udata.tests.helpers import assert_json_equal, assert_equal_dates
